+++ feature_engine/preprocess.py
@@ -1,0 +1,6 @@
+import re
+
+def preprocess_text(text):
+
+	# preprocess to remove special characters
+	return(re.sub('[^a-zA-Z ]+', '', text).lower())
